# docs/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: This is a rule engine server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rule Engine API
  version: "1.0"
paths:
...
  /api/v1/triggers:
    get:
      description: Get a list of all triggers with optional pagination.
      parameters:
        - description: Limit number of triggers returned
          in: query
          name: limit
          schema:
            type: integer
        - description: Offset for pagination
          in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: List all triggers
      tags:
        - triggers
    post:
      description: Create a new trigger for a rule.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.CreateTriggerRequest"
        description: Trigger to create
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.TriggerInfo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Create a new trigger
      tags:
        - triggers
  "/api/v1/triggers/{id}":
    delete:
      description: Delete a trigger by its unique ID.
      parameters:
        - description: Trigger ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Delete a trigger
      tags:
        - triggers
    get:
      description: Get a single trigger by its unique ID.
      parameters:
        - description: Trigger ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.TriggerInfo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Get a trigger by ID
      tags:
        - triggers
    patch:
      description: Update an existing trigger using a JSON Patch.
      parameters:
        - description: Trigger ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        $ref: "#/components/requestBodies/api.PatchOperationArray"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.TriggerInfo"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIErrorResponse"
      summary: Update a trigger
      tags:
        - triggers
...
servers:
  - url: //localhost:8080/api/v1
components:
  requestBodies:
    api.PatchOperationArray:
      content:
        application/json:
          schema:
            items:
              $ref: "#/components/schemas/api.PatchOperation"
            type: array
      description: JSON Patch operations
      required: true
  schemas:
    api.APIErrorResponse:
      properties:
        error:
          example: Error message
          type: string
      type: object
...
    api.CreateTriggerRequest:
      properties:
        condition_script:
          example: if event.device_id == 'sensor_1' then return true end
          type: string
        enabled:
          example: true
          type: boolean
        rule_id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        type:
          enum:
            - CONDITIONAL
            - CRON
          example: CONDITIONAL
          type: string
      required:
        - condition_script
        - rule_id
        - type
      type: object
...
    api.PatchOperation:
      properties:
        op:
          enum:
            - add
            - remove
            - replace
            - test
          example: replace
          type: string
        path:
          example: /name
          type: string
        value: {}
      required:
        - op
        - path
      type: object
...
    api.TriggerInfo:
      properties:
        condition_script:
          type: string
        created_at:
          type: string
        enabled:
          type: boolean
        id:
          type: string
        rule_id:
          type: string
        type:
          type: string
        updated_at:
          type: string
      type: object